Split up timeseries (#985)
* move PR aggregation to custom_aggregations

* moving vaccine backfills and region aggregation

* splitting up tests some more

* Fixing tests

* Rename some missing region_aggregation imports

* Add missing import

* One lint lets go
