+++ libs/datasets/statistical_areas.py
@@ -4,11 +4,10 @@ from typing import Mapping
 import pandas as pd
 
 from libs import pipeline
-from libs.datasets import timeseries
 from libs.datasets.timeseries import MultiRegionDataset
 from covidactnow.datapublic.common_fields import CommonFields
 from libs.datasets import dataset_utils
-
+from libs.datasets import region_aggregation
 
 CBSA_LIST_PATH = "data/census-msa/list1_2020.xls"
 
@@ -25,8 +24,8 @@ class CountyToCBSAAggregator:
     cbsa_title_map: Mapping[str, str]
 
     aggregations: List[
-        timeseries.StaticWeightedAverageAggregation
-    ] = timeseries.WEIGHTED_AGGREGATIONS
+        region_aggregation.StaticWeightedAverageAggregation
+    ] = region_aggregation.WEIGHTED_AGGREGATIONS
 
     def aggregate(
         self, dataset_in: MultiRegionDataset, reporting_ratio_required_to_aggregate=None
@@ -37,7 +36,7 @@ class CountyToCBSAAggregator:
             for fips, cbsa_code in self.county_map.items()
         }
 
-        return timeseries.aggregate_regions(
+        return region_aggregation.aggregate_regions(
             dataset_in,
             region_map,
             self.aggregations,
